ft: Navbar account dropdown menu
